# Use DataView instead of Buffer in hand-written benchmark