--- a/benchmark/package.json
+++ b/benchmark/package.json
@@ -1,8 +1,9 @@
 {
   "dependencies": {
     "benchmark": "^2.1.4",
-    "binparse": "1.2.1",
-    "destruct-js": "^0.2.9",
+    "binparse": "^2.1.0",
+    "destruct-js": "^0.2.13",
     "structron": "^0.4.3"
-  }
+  },
+  "type": "module"
 }