@keichi 2023-11-22
Thanks for pointing this out! I actually wasn't aware of the performance difference between DataView and Buffer. We initially used Buffer but then switched to DataView to support browser environments (https://github.com/keichi/binary-parser/pull/140)