@screeny05 2023-11-22
Sure!

I wasn't aware of this as well until i read this article by the v8 dev team:

https://v8.dev/blog/dataview